move random_circuit
